moved aws s3 to skeleton file
